Move tests to correct namespace
These tests were testing types from another namepace.
