+++ src/Microsoft.VisualStudio.ProjectSystem.Managed.UnitTests/Threading/Tasks/SequencialTaskExecutorTests.cs
@@ -1,11 +1,11 @@
 ﻿// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
+
 using System;
 using System.Collections.Generic;
-using Microsoft.VisualStudio.Threading.Tasks;
 using Xunit;
 using Task = System.Threading.Tasks.Task;
 
-namespace Microsoft.VisualStudio.ProjectSystem
+namespace Microsoft.VisualStudio.Threading.Tasks
 {
 
     [ProjectSystemTrait]
